Pruned frontmatter and redundant includes
