+++ packages/docs/src/content/about/getting-started.mdx
@@ -1,12 +1,8 @@
 ---
 title: Getting started
 description: Guides to getting started using NICE Design System
-section: about
-
-template: generalSidenav
-headerCTA:
-  href: https://codepen.io/pen?template=mKvYJO
-  text: Try it on CodePen
+template: detail
+sort: 2
 ---
 
 {{ site.product_name }} is the source for creating consistent, accessible and trustworthy services for NICE. It’s aimed at developers, designers and content editors. Please use our relevant guides below to get started:
