+++ packages/docs/src/content/about/index.mdx
@@ -1,9 +1,9 @@
 ---
 title: About the NICE Design System
+navTitle: About
 description: NICE Design System is a set of reusable patterns, documentation and guidelines for rapid product development.
-section: about
-
-template: generalSidenav
+template: detail
+sort: 1
 ---
 
 Its goal is to create a consistent, accessible digital experience across all of NICE's services.
