+++ packages/docs/src/content/about/roadmap.mdx
@@ -1,9 +1,8 @@
 ---
 title: Roadmap
 description: Where we're taking the product
-section: about
-
-template: generalSidenav
+template: detail
+sort: 3
 ---
 
 We've split our roadmap into short, medium and long term and by components, content, code and build. This is subject to change and will be updated on a regular basis.
