+++ packages/docs/src/content/components/action-banner/index.mdx
@@ -1,9 +1,8 @@
 ---
 title: Action banner
 description: Banner for highlighting, and giving context to, a call-to-action
-section: components
-
-template: component
+inpagenav: true
+template: detail
 ---
 
 Import { ActionBanner } from "@nice-digital/nds-action-banner"
