+++ packages/docs/src/content/components/breadcrumbs/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Breadcrumbs
 description: A breadcrumb trail is a secondary navigation, that displays as a series of links, situated below the header.
-section: components
-template: component
+
+template: detail
 
 ---
