+++ packages/docs/src/content/components/card/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Card
 description: A flexible component for displaying an item like news or guidance within a list
-section: components
-template: component
+
+template: detail
 
 ---
