+++ packages/docs/src/content/components/hero/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Hero
 description: Introduce a site or sub-site with impact
-section: components
-template: component
+
+template: detail
 
 ---
