+++ packages/docs/src/content/components/in-page-nav/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: In page nav
 description: Allows users to navigate to specific sections, within long pages of content, across any screen size.
-section: components
-template: component
+
+template: detail
 
 ---
