+++ packages/docs/src/content/components/index.mdx
@@ -1,7 +1,6 @@
 ---
 title: Components
 description: Once you've mastered the foundations, these components are further patterns for building user interfaces.
-section: components
-
 template: overview
+sort: 1
 ---
