+++ packages/docs/src/content/components/maintain-ratio/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Maintain ratio
 description: Embed media and maintain an aspect ratio across devices
-section: components
-template: component
+
+template: detail
 
 ---
