+++ packages/docs/src/content/components/page-header/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Page header
 description: Main heading and intro copy to introduce a page
-section: components
-template: component
+
+template: detail
 
 ---
