+++ packages/docs/src/content/components/panel/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Panel
 description: Blocks to highlight or separate content
-section: components
-template: component
+
+template: detail
 
 ---
