+++ packages/docs/src/content/components/phase-banner/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Phase banner
 description: Alpha and beta beta banners clearly identify non-live services
-section: components
-template: component
+
+template: detail
 
 ---
