+++ packages/docs/src/content/components/stacked-navigation/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Stacked navigation
 description: Navigation to support sub-pages within a section
-section: components
-template: component
+
+template: detail
 
 ---
