+++ packages/docs/src/content/components/tabs/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Tabs
 description: Content is separated into different panes, with only one pane viewable at a time
-section: components
-template: component
+
+template: detail
 
 ---
