+++ packages/docs/src/content/components/tag/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Tag
 description: Tag component used for labelling items in a list
-section: components
-template: component
+
+template: detail
 
 ---
