+++ packages/docs/src/content/foundations/buttons/index.mdx
@@ -1,9 +1,7 @@
 ---
 title: Buttons
 description: Button of different types, colours and sizes and combination of buttons
-section: foundations
-template: component
-
+template: detail
 ---
 
 Their main purpose is to trigger an action when the user clicks on them. It's important that the button label is clear about what action the button performs to the user.
