+++ packages/docs/src/content/foundations/colour/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Colour
 description: Colour palettes and usage, contrast guidelines and SASS variables
-section: foundations
-template: component
+
+template: detail
 
 ---
