+++ packages/docs/src/content/foundations/helpers/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Helpers
 description: Useful helper classes for common tasks
-section: foundations
-template: component
+
+template: detail
 
 ---
