+++ packages/docs/src/content/foundations/iconography/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Iconography
 description: Icons, their usage and guidance for creating new custom icons
-section: foundations
-template: component
+
+template: detail
 
 ---
