+++ packages/docs/src/content/foundations/index.mdx
@@ -1,7 +1,6 @@
 ---
 title: Foundations
 description: The foundations and patterns that are to be used for faster product development.
-section: foundations
-
 template: overview
+sort: 1
 ---
