+++ packages/docs/src/content/foundations/layout/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Grid & layout
 description: Use containers, grids and whitespace to build a structure and hierarchy on the page.
-section: foundations
-template: component
+
+template: detail
 
 ---
