+++ packages/docs/src/content/foundations/links/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Links
 description: Links and hover/active/visited/focus states
-section: foundations
-template: component
+
+template: detail
 
 ---
