+++ packages/docs/src/content/foundations/responsive/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Responsive
 description: Guidance for building mobile-first responsive web applications
-section: foundations
-template: component
+
+template: detail
 
 ---
