+++ packages/docs/src/content/foundations/spacing/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Spacing
 description: Spacing between elements and vertical rhythm
-section: foundations
-template: component
+
+template: detail
 
 ---
