+++ packages/docs/src/content/foundations/typography/index.mdx
@@ -1,8 +1,8 @@
 ---
 title: Typography
 description: Headings, paragraphs, featured text, quotations, lists, links
-section: foundations
-template: component
+
+template: detail
 
 ---
