+++ packages/docs/src/content/foundations/visibility/index.mdx
@@ -1,8 +1,7 @@
 ---
 title: Visibility
 description: Showing and hiding elements across breakpoints, for print, device orientations and screen readers
-section: foundations
-template: component
+template: detail
 
 ---
