+++ packages/docs/src/content/technical/index.mdx
@@ -1,7 +1,6 @@
 ---
 title: Technical
 description: Technical documentation and code style guides for creating NICE digital services
-section: technical
-
 template: overview
+sort: 1
 ---
