+++ packages/docs/src/content/technical/installation.mdx
@@ -1,6 +1,5 @@
 ---
 title: Installation
 description: Install the product to get started
-section: technical
-
+template: detail
 ---
