+++ packages/docs/src/content/technical/javascript.mdx
@@ -1,8 +1,7 @@
 ---
 title: JavaScript
 description: Install the product to get started
-section: technical
-
+template: detail
 ---
