+++ packages/docs/src/content/technical/sass-css/code-standards.mdx
@@ -1,6 +1,6 @@
 ---
 title: SASS code standards
 description: SASS code standards, naming conventions etc
-section: sass-css
-
+template: detail
+sort: 2
 ---
