+++ packages/docs/src/content/technical/sass-css/index.mdx
@@ -1,8 +1,6 @@
 ---
 title: SASS & CSS
 description: SASS documentation (mixins, functions, variables, selectors) and code style guide
-section: sass-css
-
 template: overview
 ---
