+++ packages/docs/src/content/technical/sass-css/libraries.mdx
@@ -1,6 +1,6 @@
 ---
 title: Libraries
 description: The third party libraries we use within our SASS
-section: sass-css
-
+template: detail
+sort: 5
 ---
