+++ packages/docs/src/content/technical/sass-css/sass-documentation.mdx
@@ -1,8 +1,8 @@
 ---
 title: SASS documentation
 description: Auto-generated SASS documentation from our SASSDoc comments
-section: sass-css
-
+template: detail
+sort: 3
 ---
 
 Auto-generated SASS documentation
