+++ packages/docs/src/content/technical/sass-css/sass-installation.mdx
@@ -1,6 +1,7 @@
 ---
 title: SASS installation
 description: How to install the NICE Design System SASS
-section: sass-css
-
+template: detail
+sort: 4
 ---
+
